vayojanamitra/backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
chromadb
requests
httpx
beautifulsoup4
apscheduler
python-dotenv
motor
pymongo
google-generativeai
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
openai
email-validator
reportlab
